+++ src/mirrulations_client/client_manager.py
@@ -1,11 +1,10 @@
 import mirrulations_client.document_processor as doc
 import mirrulations_client.documents_processor as docs
 from mirrulations_client.server_call_manager import ServerCallManager
+from mirrulations_client.client_health_call_manager import ClientHealthCallManager
 from mirrulations_core.api_call_manager import APICallManager
-import requests
 import json
 import time
-import logging
 import shutil
 import tempfile
 from pathlib import Path
@@ -16,19 +15,6 @@ API_KEY = config.read_value('CLIENT', 'API_KEY')
 CLIENT_ID = config.read_value('CLIENT', 'CLIENT_ID')
 SERVER_ADDRESS = config.read_value('CLIENT', 'SERVER_ADDRESS')
 VERSION = "v1.3"
-
-
-class ClientHealthCallManager:
-
-    def __init__(self):
-        self.base_url = 'https://hc-ping.com/457a1034-83d4-4a62-8b69-c71060db3a08'
-
-    def make_call(self):
-        requests.get(self.base_url)
-
-    def make_fail_call(self):
-        requests.get(self.base_url + '/fail')
-
 
 api_manager = APICallManager(API_KEY)
 server_manager = ServerCallManager(CLIENT_ID, SERVER_ADDRESS)
